fixup! Create url tree without currentUrlTree
